pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = ""
authors = [
    {name = "simon.yang.sh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "uiautomator2 (>=3.3.1,<4.0.0)",
    "uiautodev (>=0.9.0,<0.10.0)",
    "behave (>=1.2.6,<2.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "allure-pytest (>=2.14.3,<3.0.0)",
    "pytest-bdd (>=8.1.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
